requirements.txt
alabaster==0.7.12
arrow==1.1.1
Babel==2.9.1
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cookiecutter==1.7.3
docutils==0.16
EasyProcess==0.3
entrypoint2==0.2.4
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
packaging==21.0
patool==1.12
Pillow==8.3.1
poyo==0.5.0
Pygments==2.9.0
pyparsing==2.4.7
PyPDF4==1.27.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
pyunpack==0.2.2
recommonmark==0.7.1
requests==2.26.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
text-unidecode==1.3
urllib3==1.26.6
wincertstore==0.2
